@@ -1,23 +1,27 @@
 {
-    "name": "knpuniversity/oauth2-client-bundle",
-    "type": "symfony-bundle",
+    "name": "willy68/psr-oauth2-client",
+    "type": "psr-oauth2-client",
     "description": "Integration with league/oauth2-client to provide services",
     "keywords": ["oauth", "oauth2"],
     "homepage": "https://symfonycasts.com",
     "license": "MIT",
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "authors": [
         {
             "name": "Ryan Weaver",
             "email": "ryan@symfonycasts.com"
+        },
+        {
+            "name": "William Lety",
+            "email": "william.lety@gmail.com"
         }
     ],
     "require": {
-        "php": ">=7.2.5",
-        "symfony/framework-bundle": "^4.4|^5.0|^6.0",
-        "symfony/dependency-injection": "^4.4|^5.0|^6.0",
-        "symfony/routing": "^4.4|^5.0|^6.0",
-        "symfony/http-foundation": "^4.4|^5.0|^6.0",
+        "php": ">=8.0",
+        "psr/container": "^2.0.2",
+        "psr/http-message": "^1.0.1",
+        "psr/http-factory": "^1.0.1",
+        "mezzio/mezzio-session": "^1.12.0",
         "league/oauth2-client": "^2.0"
     },
     "autoload": {
@@ -28,13 +32,7 @@
     },
     "require-dev": {
         "league/oauth2-facebook": "^1.1|^2.0",
-        "symfony/phpunit-bridge": "^5.3.1|^6.0",
-        "symfony/security-guard": "^4.4|^5.0|^6.0",
-        "symfony/yaml": "^4.4|^5.0|^6.0",
-        "phpstan/phpstan": "^0.12"
-    },
-    "suggest": {
-        "symfony/security-guard": "For integration with Symfony's Guard Security layer"
+        "phpstan/phpstan": "^0.12.100"
     },
     "scripts": {
         "test": "simple-phpunit && phpstan analyze"
